ECV-658 chore: Moves some tests into the correct location
