+++ src/EPR.Calculator.API.UnitTests/Validators/CalculatorRunStatusDataValidatorTest.cs
@@ -5,7 +5,7 @@ using EPR.Calculator.API.Enums;
 using EPR.Calculator.API.UnitTests.TestHelpers.Fixtures;
 using EPR.Calculator.API.Validators;
 
-namespace EPR.Calculator.API.UnitTests.Validator
+namespace EPR.Calculator.API.UnitTests.Validators
 {
     [TestClass]
     public class CalculatorRunStatusDataValidatorTest
