+++ src/EPR.Calculator.API.UnitTests/Validators/ProduceBillingInstructionRequestDtoValidatorTests.cs
@@ -2,7 +2,7 @@
 using EPR.Calculator.API.Enums;
 using EPR.Calculator.API.Validators;
 
-namespace EPR.Calculator.API.UnitTests.Validator
+namespace EPR.Calculator.API.UnitTests.Validators
 {
     [TestClass]
     public class ProduceBillingInstructionRequestDtoValidatorTests
